clean up diffs.go
