+++ consensus/diffs.go
@@ -1,46 +1,52 @@
 package consensus
 
-// A block is composed of many transactions. Blocks that have transactions that
-// depend on other transactions, but the transactions must all be applied in a
-// deterministic order. Transactions cannot have inter-dependencies, meaning
-// that an output cannot be created and then spent in the same transaction. As
-// far as diffs are concenred, this means that the elements of a transaction
-// diff should be able to be applied in any order and the outcome should be the
-// same. The elements of a block diff however must be applied in a specific
-// order, as transactions may depend on each other.
+// Everything in the consensus set has a binary existence. Either it exists or
+// it doesn't, and anything that exists with the same id in the same fork will
+// have the same contents. This means that all diffs either add an element to
+// the consensus set, or they remove an element from the consensus set, and
+// that all diffs are easily reversible. The `New` flag indicates whether the
+// item is being added or removed from the consensus set. When apply a diff,
+// use the new flag as intended. When reversing a diff, flip the new flag.
 
-// An OutputDiff indicates an output that has either been added to or removed
-// from the unspent outputs set. New=true means that the output was added when
-// the block was applied, and new=false means that the output was deleted when
-// the block was applied.
+// A SiacoinOutputDiff indicates the addition or removal of a SiacoinOutput in
+// the consensus set.
 type SiacoinOutputDiff struct {
 	New           bool
 	ID            SiacoinOutputID
 	SiacoinOutput SiacoinOutput
 }
 
+// A FileContractDiff indicates the addition or removal of a FileContract in
+// the consensus set.
 type FileContractDiff struct {
 	New          bool
 	ID           FileContractID
 	FileContract FileContract
 }
 
+// A SiafundOutputDiff indicates the addition or removal of a SiafundOutput in
+// the consensus set.
 type SiafundOutputDiff struct {
 	New           bool
 	ID            SiafundOutputID
 	SiafundOutput SiafundOutput
 }
 
+// A SiafundPool diff contains the value of the siafundPool before the block
+// was applied, and after the block was applied. When applying the diff, set
+// siafundPool to `Adjusted`. When reversing the diff, set siafundPool to
+// `Previous`.
 type SiafundPoolDiff struct {
 	Previous Currency
 	Adjusted Currency
 }
 
-// commitOutputDiff takes an output diff and applies it to the state. Forward
-// indicates the direction of the blockchain.
-func (s *State) commitSiacoinOutputDiff(scod SiacoinOutputDiff, forward bool) {
+// commitSiacoinOutputDiff takes a SiacoinOutputDiff and applies it to the
+// consensus set. `applied` indicates whether the diff is being applied or
+// reversed.
+func (s *State) commitSiacoinOutputDiff(scod SiacoinOutputDiff, applied bool) {
 	add := scod.New
-	if !forward {
+	if !applied {
 		add = !add
 	}
 
@@ -67,11 +73,12 @@ func (s *State) commitSiacoinOutputDiff(scod SiacoinOutputDiff, forward bool) {
 	}
 }
 
-// commitContractDiff takes a contract diff and applies it to the state. Forward
-// indicates the direction of the blockchain.
-func (s *State) commitFileContractDiff(fcd FileContractDiff, forward bool) {
+// commitFileContractDiff takes a FileContractDiff and applies it to the
+// consensus set.  `applied` indicates whether the diff is being applied or
+// reversed.
+func (s *State) commitFileContractDiff(fcd FileContractDiff, applied bool) {
 	add := fcd.New
-	if !forward {
+	if !applied {
 		add = !add
 	}
 
@@ -98,9 +105,12 @@ func (s *State) commitFileContractDiff(fcd FileContractDiff, forward bool) {
 	}
 }
 
-func (s *State) commitSiafundOutputDiff(sfod SiafundOutputDiff, forward bool) {
+// commitSiafundOutputDiff takes a SiafundOutputDiff and applies it to the
+// consensus set. `applied` indicates whether the diff is being applied or
+// reversed.
+func (s *State) commitSiafundOutputDiff(sfod SiafundOutputDiff, applied bool) {
 	add := sfod.New
-	if !forward {
+	if !applied {
 		add = !add
 	}
 
@@ -127,25 +137,33 @@ func (s *State) commitSiafundOutputDiff(sfod SiafundOutputDiff, forward bool) {
 	}
 }
 
-func (s *State) commitSiafundPoolDiff(sfpd SiafundPoolDiff, forward bool) {
-	if forward {
+// commitSiafundPoolDiff takes a SiafundPoolDiff and applies it to the
+// consensus set. `applied` indicates whether the diff is being applied or
+// reversed.
+func (s *State) commitSiafundPoolDiff(sfpd SiafundPoolDiff, applied bool) {
+	if applied {
 		s.siafundPool = sfpd.Adjusted
 	} else {
 		s.siafundPool = sfpd.Previous
 	}
 }
 
-func (s *State) applyDiffSet(bn *blockNode, direction bool) {
-	// Sanity check - diffs should have already been generated for this node.
+// applyDiffSet takes all of the diffs in a block and applies them to the
+// consensus set. `applied` indicates whether the diffs are being applied or
+// reversed. The ordering is important, because transactions within the same
+// block are allowed to depend on each other. When reversing diffs, they must
+// be reversed in the opposite order that they were applied.
+func (s *State) applyDiffSet(bn *blockNode, applied bool) {
+	// Sanity check
 	if DEBUG {
+		// Diffs should have already been generated for this node.
 		if !bn.diffsGenerated {
 			panic("misuse of applyDiffSet - diffs have not been generated!")
 		}
-	}
-	// Sanity check - current node must be the input node's parent if direction
-	// = true, and current node must be the input node if direction = false.
-	if DEBUG {
-		if direction {
+
+		// Current node must be the input node's parent if applied = true, and
+		// current node must be the input node if applied = false.
+		if applied {
 			if bn.parent.block.ID() != s.currentBlockID {
 				panic("applying a block node when it's not a valid successor")
 			}
@@ -156,79 +174,59 @@ func (s *State) applyDiffSet(bn *blockNode, direction bool) {
 		}
 	}
 
-	// Apply the siacoin, file contract, and siafund diffs.
-	s.commitSiafundPoolDiff(bn.siafundPoolDiff, direction)
-	for _, scod := range bn.siacoinOutputDiffs {
-		s.commitSiacoinOutputDiff(scod, direction)
-	}
-	for _, fcd := range bn.fileContractDiffs {
-		s.commitFileContractDiff(fcd, direction)
-	}
-	for _, sfod := range bn.siafundOutputDiffs {
-		s.commitSiafundOutputDiff(sfod, direction)
-	}
+	// The sets of diffs can be applied and reversed in any order, so the
+	// ordering is kept when applying and reversing sets. However, within the
+	// sets, the diffs must be reversed in the opposite order that they were
+	// applied, due to transactions being able to depend on each other. This
+	// results is messier for loops when applied is false.
+	if applied {
+		for _, scod := range bn.siacoinOutputDiffs {
+			s.commitSiacoinOutputDiff(scod, applied)
+		}
+		for _, fcd := range bn.fileContractDiffs {
+			s.commitFileContractDiff(fcd, applied)
+		}
+		for _, sfod := range bn.siafundOutputDiffs {
+			s.commitSiafundOutputDiff(sfod, applied)
+		}
 
-	// Manage the delayed outputs that have been created by the node.
-	if direction {
-		s.delayedSiacoinOutputs[bn.height] = bn.delayedSiacoinOutputs
-	} else {
-		delete(s.delayedSiacoinOutputs, bn.height)
-	}
-
-	// Update the current path and currentBlockID
-	if direction {
+		s.commitSiafundPoolDiff(bn.siafundPoolDiff, applied)
 		s.currentBlockID = bn.block.ID()
 		s.currentPath[bn.height] = bn.block.ID()
+		s.delayedSiacoinOutputs[bn.height] = bn.delayedSiacoinOutputs
 	} else {
-		delete(s.currentPath, bn.height)
+		for i := len(bn.siacoinOutputDiffs) - 1; i >= 0; i-- {
+			s.commitSiacoinOutputDiff(bn.siacoinOutputDiffs[i], applied)
+		}
+		for i := len(bn.fileContractDiffs) - 1; i >= 0; i-- {
+			s.commitFileContractDiff(bn.fileContractDiffs[i], applied)
+		}
+		for i := len(bn.siafundOutputDiffs) - 1; i >= 0; i-- {
+			s.commitSiafundOutputDiff(bn.siafundOutputDiffs[i], applied)
+		}
+
+		s.commitSiafundPoolDiff(bn.siafundPoolDiff, applied)
 		s.currentBlockID = bn.parent.block.ID()
-	}
-}
-
-// applyMinerSubsidy adds all of the outputs recorded in the MinerPayouts to
-// the state, and returns the corresponding set of diffs.
-func (s *State) applyMinerSubsidy(bn *blockNode) {
-	for i, payout := range bn.block.MinerPayouts {
-		id := bn.block.MinerPayoutID(i)
-		s.delayedSiacoinOutputs[s.height()][id] = payout
-		bn.delayedSiacoinOutputs[id] = payout
-	}
-	return
-}
-
-// applyDelayedSiacoinOutputMaintenance goes through all of the outputs that
-// have matured and adds them to the list of siacoinOutputs.
-func (s *State) applyDelayedSiacoinOutputMaintenance(bn *blockNode) {
-	for id, sco := range s.delayedSiacoinOutputs[bn.height-MaturityDelay] {
-		// Sanity check - the output should not already be in the
-		// siacoinOuptuts list.
-		if DEBUG {
-			_, exists := s.siacoinOutputs[id]
-			if exists {
-				panic("trying to add a delayed output when the output is already there")
-			}
-		}
-		s.siacoinOutputs[id] = sco
-
-		// Create and add the diff.
-		scod := SiacoinOutputDiff{
-			New:           true,
-			ID:            id,
-			SiacoinOutput: sco,
-		}
-		bn.siacoinOutputDiffs = append(bn.siacoinOutputDiffs, scod)
+		delete(s.currentPath, bn.height)
+		delete(s.delayedSiacoinOutputs, bn.height)
 	}
 }
 
 // generateAndApplyDiff will verify the block and then integrate it into the
-// consensus state.
+// consensus state. These two actions must happen at the same time because
+// transactions are allowed to depend on each other. We can't be sure that a
+// transaction is valid unless we have applied all of the previous transactions
+// in the block, which means we need to apply while we verify.
 func (s *State) generateAndApplyDiff(bn *blockNode) (err error) {
-	// Sanity check - generate should only be called if the diffs have not yet
-	// been generated - current node must be the input node's parent.
+	// Sanity check
 	if DEBUG {
+		// Generate should only be called if the diffs have not yet been
+		// generated.
 		if bn.diffsGenerated {
 			panic("misuse of generateAndApplyDiff")
 		}
+
+		// Current node must be the input node's parent.
 		if bn.parent.block.ID() != s.currentBlockID {
 			panic("applying a block node when it's not a valid successor")
 		}
@@ -254,21 +252,19 @@ func (s *State) generateAndApplyDiff(bn *blockNode) (err error) {
 	for _, txn := range bn.block.Transactions {
 		err = s.validTransaction(txn)
 		if err != nil {
-			direction := false // set direction to false because the diffs are being removed.
-			s.applyDiffSet(bn, direction)
+			applied := false // set applied to false because the diffs are being removed.
+			s.applyDiffSet(bn, applied)
 			return
 		}
 
 		s.applyTransaction(bn, txn)
 	}
 
-	// Perform maintanence on all open contracts.
-	s.applyContractMaintenance(bn)
-	s.applyDelayedSiacoinOutputMaintenance(bn)
-	s.applyMinerSubsidy(bn)
+	// Perform maintanence on the block.
+	s.applyMaintenance(bn)
 
-	// The final diff to be applied is to mark what the ending siafundPool
-	// balance is.
+	// The final thing is to update the siafundPoolDiff to indicate where the
+	// siafund pool ended up.
 	bn.siafundPoolDiff.Adjusted = s.siafundPool
 
 	return
